python/requirements.txt
Flask
gunicorn
pyhdb
python_jwt
pycrypto
jws